+++ Source/Rhetos.Configuration.Autofac/IRhetosHostBuilder.cs
@@ -17,23 +17,40 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Text;
 using Autofac;
 using Rhetos.Logging;
 using Rhetos.Utilities;
+using System;
+using System.Collections.Generic;
 
 namespace Rhetos
 {
     public interface IRhetosHostBuilder
     {
         IRhetosHostBuilder UseBuilderLogProvider(ILogProvider logProvider);
+
         IRhetosHostBuilder ConfigureConfiguration(Action<IConfigurationBuilder> configureAction);
+
+        /// <summary>
+        /// Configures services and plugins in the dependency injection container.
+        /// </summary>
         IRhetosHostBuilder ConfigureContainer(Action<ContainerBuilder> configureAction);
+
         IRhetosHostBuilder UseCustomContainerConfiguration(Action<IConfiguration, ContainerBuilder, List<Action<ContainerBuilder>>> containerConfigurationAction);
+
+        /// <summary>
+        /// This method is intended for external applications that reference the generated Rhetos application
+        /// (CLI utilities, for example) to specify the Rhetos app root folder.
+        /// </summary>
+        /// <remarks>
+        /// In most cases, the root folder is automatically configured, and there is no need to call this
+        /// method directly from the custom application code.
+        /// </remarks>
         IRhetosHostBuilder UseRootFolder(string rootFolder);
+
+        /// <summary>
+        /// Create the <see cref="RhetosHost"/> instance that was previously configured by other <see cref="IRhetosHostBuilder"/> methods.
+        /// </summary>
         RhetosHost Build();
     }
 }